+++ src/bk-user/bkuser/biz/organization.py
@@ -16,10 +16,11 @@
 # to the current version of the project delivered to anyone in the future.
 
 import datetime
-from typing import Dict, List
+from collections import defaultdict
+from typing import Dict, List, Set
 
 from django.db import transaction
-from django.db.models import QuerySet
+from django.db.models import Prefetch, Q, QuerySet
 from django.utils import timezone
 
 from bkuser.apps.data_source.models import (
@@ -29,7 +30,7 @@ from bkuser.apps.data_source.models import (
     DataSourceUserDeprecatedPasswordRecord,
     LocalDataSourceIdentityInfo,
 )
-from bkuser.apps.tenant.models import TenantDepartment
+from bkuser.apps.tenant.models import TenantDepartment, TenantUser
 from bkuser.common.constants import PERMANENT_TIME
 from bkuser.common.hashers import make_password
 
@@ -177,3 +178,111 @@ class TenantDepartmentHandler:
             }
             for tenant_dept in tenant_depts
         }
+
+
+class TenantOrgPathHandler:
+    @staticmethod
+    def get_dept_organization_path_map(tenant_depts: QuerySet[TenantDepartment]) -> Dict[int, str]:
+        """获取租户部门的组织路径信息"""
+        data_source_dept_ids = {dept.data_source_department_id for dept in tenant_depts}
+
+        # 获取部门关系及所有需要的祖先节点
+        relations = TenantOrgPathHandler._get_relations_with_ancestors(data_source_dept_ids)
+        # 数据源部门 ID -> 组织路径
+        org_path_map = TenantOrgPathHandler._build_org_path_map(relations, include_self=False)
+
+        # 租户部门 ID -> 组织路径
+        return {
+            dept.id: org_path_map.get(dept.data_source_department_id, dept.data_source_department.name)
+            for dept in tenant_depts
+        }
+
+    @staticmethod
+    def get_user_organization_paths_map(tenant_users: QuerySet[TenantUser]) -> Dict[str, List[str]]:
+        """获取租户用户的组织路径信息"""
+        data_source_user_ids = [tenant_user.data_source_user_id for tenant_user in tenant_users]
+
+        # 数据源用户 ID -> [数据源部门 ID1， 数据源部门 ID2]
+        user_dept_id_map = defaultdict(list)
+        for relation in DataSourceDepartmentUserRelation.objects.filter(user_id__in=data_source_user_ids):
+            user_dept_id_map[relation.user_id].append(relation.department_id)
+
+        # 数据源部门 ID 集合
+        data_source_dept_ids: Set[int] = set().union(*user_dept_id_map.values())
+
+        # 获取部门关系及所有需要的祖先节点
+        relations = TenantOrgPathHandler._get_relations_with_ancestors(data_source_dept_ids)
+        # 数据源部门 ID -> 组织路径
+        org_path_map = TenantOrgPathHandler._build_org_path_map(relations)
+
+        # 租户用户 ID -> 组织路径列表
+        return {
+            user.id: [
+                org_path_map[dept_id]
+                for dept_id in user_dept_id_map[user.data_source_user_id]
+                if dept_id in org_path_map
+            ]
+            for user in tenant_users
+        }
+
+    @staticmethod
+    def _get_relations_with_ancestors(data_source_dept_ids: Set[int]) -> QuerySet[DataSourceDepartmentRelation]:
+        """获取部门关系及预取所有需要的祖先节点"""
+        # 按照左值排序，确保祖先节点在后续查询中已经加载
+        relations = DataSourceDepartmentRelation.objects.select_related("department").filter(
+            department_id__in=data_source_dept_ids
+        )
+
+        # 按照 tree_id 分组，获取每个 tree 的最大左值和最小右值
+        tree_groups: Dict[int, Dict[str, float]] = defaultdict(lambda: {"max_lft": 0, "min_rght": float("inf")})
+        for rel in relations:
+            tree_groups[rel.tree_id]["max_lft"] = max(tree_groups[rel.tree_id]["max_lft"], rel.lft)
+            tree_groups[rel.tree_id]["min_rght"] = min(tree_groups[rel.tree_id]["min_rght"], rel.rght)
+
+        # 构造组合查询条件
+        # 覆盖每个树中节点所涉及到的所有祖先节点（左值小于等于当前节点左值，右值大于等于当前节点右值）
+        combined_query = Q()
+        for tree_id, group in tree_groups.items():
+            combined_query |= Q(lft__lte=group["max_lft"], rght__gte=group["min_rght"], tree_id=tree_id)
+
+        # 批量预取父部门（单次查询完成所有祖先部门的获取）
+        return relations.prefetch_related(
+            Prefetch(
+                "parent",
+                queryset=DataSourceDepartmentRelation.objects.select_related("department").filter(combined_query),
+            )
+        )
+
+    @staticmethod
+    def _build_org_path_map(
+        relations: QuerySet[DataSourceDepartmentRelation], include_self: bool = True
+    ) -> Dict[int, str]:
+        """构建部门 ID -> 组织路径映射"""
+        # 数据源部门 ID -> 路径片段
+        path_cache: Dict[int, List] = {}
+        # 数据源部门 ID -> 组织路径
+        org_path_map: Dict[int, str] = {}
+
+        for rel in relations:
+            # 如果父部门已经计算过路径，直接拼接
+            if rel.parent and rel.parent.department_id in path_cache:
+                current_path = path_cache[rel.parent.department_id] + [rel.department.name]
+            else:
+                current_path = [rel.department.name]
+                current_parent = rel.parent
+                # 递归获取所有祖先节点
+                while current_parent:
+                    current_path.append(current_parent.department.name)
+                    current_parent = current_parent.parent
+
+                # 反转路径，使得路径从根节点到当前节点
+                current_path = list(reversed(current_path))
+
+            path_cache[rel.department_id] = current_path
+
+            # 若 include_self 为 False，则不包含当前节点
+            org_path_map[rel.department_id] = "/".join(
+                current_path[: None if include_self else -1] if current_path else []
+            )
+
+        return org_path_map
